refactor(MFLP-11): Refactor the Add to Cart
- add static method
