+++ src/main/java/api/buyhood/domain/cart/dto/response/CartRes.java
@@ -1,15 +1,22 @@
 package api.buyhood.domain.cart.dto.response;
 
-import api.buyhood.domain.cart.entity.Cart;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.util.List;
+
 @Getter
 public class CartRes {
-    private final Cart cart;
+    private final List<CartItemRes> cartItemRes;
 
     @Builder
-    private CartRes(Cart cart) {
-        this.cart = cart;
+    private CartRes(List<CartItemRes> cartItemRes) {
+        this.cartItemRes = cartItemRes;
+    }
+
+    public static CartRes of(List<CartItemRes> cartItemRes) {
+        return CartRes.builder()
+                .cartItemRes(cartItemRes)
+                .build();
     }
 }
