+++ src/main/java/api/buyhood/domain/cart/service/CartService.java
@@ -1,6 +1,7 @@
 package api.buyhood.domain.cart.service;
 
 import api.buyhood.domain.cart.dto.request.CreateCartReq;
+import api.buyhood.domain.cart.dto.response.CartItemRes;
 import api.buyhood.domain.cart.dto.response.CartRes;
 import api.buyhood.domain.cart.entity.Cart;
 import api.buyhood.domain.cart.entity.CartItem;
@@ -26,21 +27,23 @@ public class CartService {
         //todo: product 존재 여부 확인 추가하기
         List<CartItem> cartItemList = createCartReq.getCartItems().stream()
                 .map(item ->
-                        CartItem.builder()
-                                .productId(item.getProductId())
-                                .quantity(item.getQuantity())
-                                .build()
+                        CartItem.of(item.getProductId(), item.getQuantity())
                 )
                 .toList();
 
-        Cart cart = Cart.builder()
-                .cart(cartItemList)
-                .build();
-
+        Cart cart = Cart.of(cartItemList);
         cartRepository.add(userId, cart);
 
-        return CartRes.builder()
-                .cart(cart)
-                .build();
+        return getCartRes(cart);
+    }
+
+    //dto 변환
+    private CartRes getCartRes(Cart cart) {
+        List<CartItemRes> cartList = cart.getCart().stream()
+                .map(cartItem ->
+                        CartItemRes.of(cartItem.getProductId(), cartItem.getQuantity())
+                ).toList();
+
+        return CartRes.of(cartList);
     }
 }
